sdc: new flag ELONGNAME for longer entry names
If an entry was named, it was limited to a 64 byte string (including
the null byte). Now, if the SDC_ELONGNAME flag is present in the entry
flags, the max string size including the null byte is 512 bytes. This
is for tiny implementations of sdc, which might only support the regular
name size of 64 bytes.